Reorg of FTDI Touch UI files; correct build errors under certain configs
- Reorganized some files
- Fixed compilation and link errors under certain configurations
