+++ Marlin/src/lcd/extui/lib/ftdi_eve_touch_ui/screens/bio_printing_dialog_box.cpp
@@ -25,8 +25,6 @@
 
 #ifdef FTDI_BIO_PRINTING_DIALOG_BOX
 
-#include "../ftdi_eve_lib/extras/circular_progress.h"
-
 using namespace FTDI;
 using namespace ExtUI;
 using namespace Theme;
