+++ Marlin/src/lcd/extui/lib/ftdi_eve_touch_ui/screens/cocoa_press_status_screen.cpp
@@ -26,8 +26,6 @@
 
 #ifdef FTDI_COCOA_STATUS_SCREEN
 
-#include "../ftdi_eve_lib/extras/poly_ui.h"
-
 #include "cocoa_press_ui.h"
 
 #define POLY(A) PolyUI::poly_reader_t(A, sizeof(A)/sizeof(A[0]))
